Update calculator.py
removed float

+++ calculator.py
@@ -2,34 +2,34 @@ import re
 
 def calculator(number1, number2, operator):
 	# adding floats to number 1 and 2
-    number1=float(number1)
-    number2=float(number2)
+    number1=(number1)
+    number2=(number2)
 	# python operators for the calculator
     if(operator=="+"):
-        return(float(number1+number2))
+        return(number1+number2)
     if(operator=="-"):
-        return(float(number1-number2))
+        return(number1-number2)
     if(operator=="*"):
-        return(float(number1*number2))
+        return(number1*number2)
     if(operator=="/"):
         return(number1/number2)
     if(operator=="//"):
         return(number1//number2)
     if(operator=="**"):
-        return(float(number1**number2))
+        return(number1**number2)
     else:
         return False
 
 def parse_input():
+	# ask user for input
 	ParseThis=input("Enter equation: ")
+	# split the string 
 	first_num = ParseThis.split()[0]
 	second_num = ParseThis.split()[2]
 	opera = ParseThis.split()[1]
 	Num1= first_num
 	Num2= second_num
 	symbol = opera
-	
-	
-	#In-num=re.findall(r'[0-9\.]+|[^0-9\.]+', ParseThis)
+	# sends the calculator the numbers and operator
 	print(calculator(Num1, symbol, Num2))
